fix(trigger): handle root and case in ast builder
