+++ packages/app-builder/src/components/AstBuilder/Root/index.ts
@@ -1,1 +1,2 @@
+export * from './RootAnd';
 export * from './RootOrWithAnd';
